handlers/parser/requirements.txt
requests
spacy
pdfminer.six
boto3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz